+++ weixin-java-mp/src/test/java/me/chanjar/weixin/mp/demo/WxMpDemoServer.java
@@ -5,7 +5,7 @@ import me.chanjar.weixin.mp.api.WxMpConfigStorage;
 import me.chanjar.weixin.mp.api.WxMpMessageHandler;
 import me.chanjar.weixin.mp.api.WxMpMessageRouter;
 import me.chanjar.weixin.mp.api.WxMpService;
-import me.chanjar.weixin.mp.api.impl.apache.WxMpServiceImpl;
+
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.servlet.ServletHandler;
 import org.eclipse.jetty.servlet.ServletHolder;
@@ -47,7 +47,7 @@ public class WxMpDemoServer {
         .fromXml(is1);
 
       wxMpConfigStorage = config;
-      wxMpService = new WxMpServiceImpl();
+      wxMpService = new WxMpServiceImpl0();
       wxMpService.setWxMpConfigStorage(config);
 
       WxMpMessageHandler logHandler = new DemoLogHandler();
