+++ weixin-java-pay/pom.xml
@@ -19,13 +19,6 @@
       <artifactId>weixin-java-common</artifactId>
       <version>${project.version}</version>
     </dependency>
-
-    <dependency>
-      <groupId>org.jodd</groupId>
-      <artifactId>jodd-http</artifactId>
-      <version>3.7</version>
-      <!-- 由于较新的3.8版本需要jdk8，故而此处采用较低版本 -->
-    </dependency>
     <dependency>
       <groupId>com.github.binarywang</groupId>
       <artifactId>qrcode-utils</artifactId>
